# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/staff": {
      "post": {
        "operationId": "StaffController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an internal staff member",
        "tags": [
          "staff"
        ]
      },
      "get": {
        "operationId": "StaffController_list_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive match on username or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_ADMIN",
                "ADMIN",
                "EDITOR",
                "VIEWER",
                "USER"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List internal staff (paginated, searchable, role-filterable)",
        "tags": [
          "staff"
        ]
      }
    },
    "/v1/admin/staff/{id}": {
      "get": {
        "operationId": "StaffController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one staff member",
        "tags": [
          "staff"
        ]
      },
      "delete": {
        "operationId": "StaffController_deactivate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a staff member",
        "tags": [
          "staff"
        ]
      }
    },
    "/v1/admin/staff/{id}/roles": {
      "patch": {
        "operationId": "StaffController_replaceRoles_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a staff member roles",
        "tags": [
          "staff"
        ]
      }
    },
...
  },
  "info": {
    "title": "Groundwork API",
    "description": "Fullstack boilerplate API — authentication, RBAC, file uploads, audit log, and transactional email.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStaffDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "staff@acme.example"
          },
          "username": {
            "type": "string",
            "example": "Jane Agent"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 8
          },
          "roles": {
            "type": "array",
            "example": [
              "EDITOR"
            ],
            "items": {
              "type": "string",
              "enum": [
                "SUPER_ADMIN",
                "ADMIN",
                "EDITOR",
                "VIEWER",
                "USER"
              ]
            }
          }
        },
        "required": [
          "email",
          "username",
          "password",
          "roles"
        ]
      },
      "UpdateStaffRolesDto": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "example": [
              "ADMIN"
            ],
            "items": {
              "type": "string",
              "enum": [
                "SUPER_ADMIN",
                "ADMIN",
                "EDITOR",
                "VIEWER",
                "USER"
              ]
            }
          }
        },
        "required": [
          "roles"
        ]
      },
...
    }
  }
}