# apps/backend/src/staff/dto/create-staff.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import {
  ArrayNotEmpty,
  IsArray,
  IsEmail,
  IsEnum,
  IsNotEmpty,
  IsString,
  MinLength,
} from 'class-validator';
import { Role } from '../../types/role.enum';

export class CreateStaffDto {
  @ApiProperty({ example: 'staff@acme.example' })
  @IsEmail()
  email: string;

  @ApiProperty({ example: 'Jane Agent' })
  @IsString()
  @IsNotEmpty()
  username: string;

  @ApiProperty({ example: 'password123', minLength: 8 })
  @IsString()
  @MinLength(8)
  password: string;

  @ApiProperty({ enum: Role, isArray: true, example: [Role.EDITOR] })
  @IsArray()
  @ArrayNotEmpty()
  @IsEnum(Role, { each: true })
  roles: Role[];
}